+++ app/src/main/java/gov/hhs/cdc/trustedintermediary/etor/EtorDomainRegistration.java
@@ -5,8 +5,8 @@ import gov.hhs.cdc.trustedintermediary.domainconnector.DomainConnector;
 import gov.hhs.cdc.trustedintermediary.domainconnector.DomainRequest;
 import gov.hhs.cdc.trustedintermediary.domainconnector.DomainResponse;
 import gov.hhs.cdc.trustedintermediary.domainconnector.HttpEndpoint;
-import gov.hhs.cdc.trustedintermediary.etor.order.PatientDemographicsController;
-import gov.hhs.cdc.trustedintermediary.etor.order.PatientDemographicsResponse;
+import gov.hhs.cdc.trustedintermediary.etor.demographics.PatientDemographicsController;
+import gov.hhs.cdc.trustedintermediary.etor.demographics.PatientDemographicsResponse;
 import gov.hhs.cdc.trustedintermediary.wrappers.Logger;
 import java.io.IOException;
 import java.io.InputStream;
