+++ app/src/main/java/gov/hhs/cdc/trustedintermediary/etor/demographics/PatientDemographicsResponse.java
@@ -1,4 +1,4 @@
-package gov.hhs.cdc.trustedintermediary.etor.order;
+package gov.hhs.cdc.trustedintermediary.etor.demographics;
 
 import javax.annotation.Nonnull;
